+++ e2e-test/cypress/integration/change_user_password_spec.js
@@ -127,4 +127,19 @@ describe("Users/Groups Dashboard", function() {
       .should("be.visible")
       .contains("Password successfully changed");
   });
+
+  it("Root can edit all user passwords", function() {
+    // Log in as root
+    cy.login("root", "root-secret");
+    cy.visit("/users");
+
+    // Check if the button is indeed visible
+    cy.get("[data-test=edit-user-auditUser]").should("be.visible");
+    cy.get("[data-test=edit-user-dviolin]").should("be.visible");
+    cy.get("[data-test=edit-user-jdoe]").should("be.visible");
+    cy.get("[data-test=edit-user-jxavier]").should("be.visible");
+    cy.get("[data-test=edit-user-mstein]").should("be.visible");
+    cy.get("[data-test=edit-user-pkleffmann]").should("be.visible");
+    cy.get("[data-test=edit-user-thouse]").should("be.visible");
+  });
 });