+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - A new user called 'root' cannot be added [#369](https://github.com/openkfw/TruBudget/issues/369)
 - If a user changes his/her password, the new password has to follow security guidelines [#370](https://github.com/openkfw/TruBudget/issues/370)
+- Allow user 'root' to change the passwords of all users[#366](https://github.com/openkfw/TruBudget/issues/366)
+- Disallow root to create projects [#375](https://github.com/openkfw/TruBudget/issues/375)
+- Allow user 'root' to change the passwords of all users[#366](https://github.com/openkfw/TruBudget/issues/366)
 
 <!-- ### Deprecated -->